requirements.txt
deepmerge
pyahocorasick
pypinyin
requests
urllib3
